fix errors, refactor milestonegroup
