+++ frontend/src/lib/components/DataDisplay/GalleryDisplay.svelte
@@ -41,6 +41,8 @@
 		const index = data.indexOf(item);
 		return componentProps[index];
 	});
+
+	$: console.log('data in gallery: ', data);
 </script>
 
 <div class="mx-auto w-full">
